initial playtest bugfixes/tweaks

+++ game/ClassLibrary1/Pieces/IRepair.cs
@@ -8,8 +8,8 @@ namespace ClassLibrary1.Pieces
 {
     public interface IRepair : IBuilder
     {
-        public double Rate { get; }
-        public double RateBase { get; }
+        public int Rate { get; }
+        public int RateBase { get; }
 
         internal void Upgrade(IRepair.Values values);
 
@@ -18,13 +18,13 @@ namespace ClassLibrary1.Pieces
         {
             public readonly IBuilder.Values Builder;
 
-            private readonly double _rate;
-            public Values(IBuilder.Values builder, double rate)
+            private readonly int _rate;
+            public Values(IBuilder.Values builder, int rate)
             {
                 this.Builder = builder;
                 this._rate = rate;
             }
-            public double Rate => _rate;
+            public int Rate => _rate;
         }
     }
 }
